+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # pending changes
 
 - ADDED open graphs metadata
-- FIXED modules count on roster page
+- FIXED roster - modules count
 - ADDED favicon
-- ADDED BMS roster section (available also from Landing page)
+- ADDED roster - BMS section (available also from Landing page)
+- ADDED roster - display module popularity (progress bar) for each module